requirements.txt
boto3
aws-cdk-lib
constructs
pydantic>=1.10,<3
